// frontend/src/api/generated/models/userDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SIFES API
 * API do Sistema do Instituto Federal de Esperantina — Diário de Aula
 * OpenAPI spec version: 1.0
 */
import type { UserAlunoResumoDto } from './userAlunoResumoDto';
import type { UserDtoRole } from './userDtoRole';
import type { UserProfessorResumoDto } from './userProfessorResumoDto';

export interface UserDto {
  id: string;
  login: string;
  role: UserDtoRole;
  precisaTrocarSenha: boolean;
  /** @nullable */
  fotoUrl: string | null;
  criadoEm: string;
  /** @nullable */
  professor: UserProfessorResumoDto | null;
  /** @nullable */
  aluno: UserAlunoResumoDto | null;
}